Added assertion to el_dist_matrix overload of detail::make_matrix()
Current implementations of detail::copy_matrix() for el_dist_matrix objects assume
that the number of variables (here: theta grid points) is equal across all MPI
processes a.k.a. all local matrices are equally-sized. This assumption does not
always hold, but this is subject to a later (more complex) bugfix.